# .generated/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Greedible API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for Greedible Backend - Restaurant Management System",
    "contact": {
      "name": "Greedible Support",
      "email": "support@greedible.com"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Use \"Bearer {token}\" format."
      }
    },
...
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current origin"
    }
  ],
  "paths": {
...
    "/api/customers/update-address": {
      "put": {
        "summary": "Update customer address",
        "description": "Updates the address for the authenticated customer. Requires valid JWT token.",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "address"
                ],
                "properties": {
                  "address": {
                    "type": "object",
                    "required": [
                      "ward",
                      "district",
                      "street",
                      "houseNumber"
                    ],
                    "properties": {
                      "ward": {
                        "type": "string",
                        "example": "Ward 5"
                      },
                      "district": {
                        "type": "string",
                        "example": "District 1"
                      },
                      "street": {
                        "type": "string",
                        "example": "Nguyen Trai"
                      },
                      "houseNumber": {
                        "type": "string",
                        "example": "123A"
                      },
                      "buildingName": {
                        "type": "string",
                        "nullable": true,
                        "example": "Sunrise Building"
                      },
                      "block": {
                        "type": "string",
                        "nullable": true,
                        "example": "Block B"
                      },
                      "floor": {
                        "type": "string",
                        "nullable": true,
                        "example": "10"
                      },
                      "roomNumber": {
                        "type": "string",
                        "nullable": true,
                        "example": "1005"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Address updated successfully"
                    },
                    "address": {
                      "type": "object",
                      "description": "Updated address object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required address fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Required address fields are missing"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No token provided or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "No token provided"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error during address update",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Error updating address"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/create": {
      "post": {
        "summary": "Create an order for authenticated customer",
        "description": "Creates a new order for authenticated customers. Supports delivery info, payment method, and loyalty points usage.",
        "tags": ["Orders"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "items",
                  "delivery_address",
                  "delivery_distance"
                ],
                "properties": {
                  "items": {
                    "type": "array",
                    "description": "Items being ordered",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "number" },
                        "price": { "type": "number" },
                        "quantity": { "type": "number" }
                      }
                    },
                    "example": [
                      {
                        "id": 11,
                        "price": 34000,
                        "quantity": 2
                      }
                    ]
                  },
                  "delivery_address": {
                    "type": "string",
                    "example": "123, Nguyen Hue, Ben Nghe Ward, District 1, Building: Sunrise, Floor 1, Room 101"
                  },
                  "delivery_distance": {
                    "type": "number",
                    "example": 5
                  },
                  "delivery_charge": {
                    "type": "number",
                    "example": 20000
                  },
                  "payment_method": {
                    "type": "string",
                    "enum": ["cash", "momo", "momo wallet"],
                    "example": "momo wallet"
                  },
                  "loyalty_points_used": {
                    "type": "number",
                    "default": 0
                  },
                  "loyalty_points_earned": {
                    "type": "number",
                    "default": 0
                  },
                  "status": {
                    "type": "string",
                    "example": "Pending"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Order created successfully",
                  "orderId": 10
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          },
          "401": {
            "description": "Unauthorized - token missing or invalid",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "No token provided"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}